pyproject.toml
[project]
name = "discord-mcp"
version = "0.1.0"
dependencies = [
   "discord.py>=2.3.0",
   "mcp>=0.1.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_mcp"]

[project.scripts]
mcp-discord = "discord_mcp:main"